refactor: 불필요한 파일 삭제 및 mockData 관련 경로 정리
